// types/graphql-types.ts
export type Maybe<T> = T | null
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K]
}
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string
  String: string
  Boolean: boolean
  Int: number
  Float: number
  /**
   * A date string, such as 2007-12-03, compliant with the ISO 8601 standard for
   * representation of dates and times using the Gregorian calendar.
   */
  Date: any
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: any
}
// ...
export type Internal = {
  content?: Maybe<Scalars["String"]>
  contentDigest: Scalars["String"]
  description?: Maybe<Scalars["String"]>
  fieldOwners?: Maybe<Array<Maybe<Scalars["String"]>>>
  ignoreType?: Maybe<Scalars["Boolean"]>
  mediaType?: Maybe<Scalars["String"]>
  owner: Scalars["String"]
  type: Scalars["String"]
}
// ...
export type Mdx = Node & {
  rawBody: Scalars["String"]
  fileAbsolutePath: Scalars["String"]
  frontmatter?: Maybe<MdxFrontmatter>
  slug?: Maybe<Scalars["String"]>
  body: Scalars["String"]
  excerpt: Scalars["String"]
  headings?: Maybe<Array<Maybe<MdxHeadingMdx>>>
  html?: Maybe<Scalars["String"]>
  mdxAST?: Maybe<Scalars["JSON"]>
  tableOfContents?: Maybe<Scalars["JSON"]>
  timeToRead?: Maybe<Scalars["Int"]>
  wordCount?: Maybe<MdxWordCount>
  id: Scalars["ID"]
  parent?: Maybe<Node>
  children: Array<Node>
  internal: Internal
}
// ...
export type MdxFrontmatter = {
  title: Scalars["String"]
  path?: Maybe<Scalars["String"]>
  date?: Maybe<Scalars["Date"]>
  description?: Maybe<Scalars["String"]>
}
// ...
export type MdxHeadingMdx = {
  value?: Maybe<Scalars["String"]>
  depth?: Maybe<Scalars["Int"]>
}

export type MdxHeadingMdxFilterInput = {
  value?: Maybe<StringQueryOperatorInput>
  depth?: Maybe<IntQueryOperatorInput>
}

export type MdxHeadingMdxFilterListInput = {
  elemMatch?: Maybe<MdxHeadingMdxFilterInput>
}

export type MdxSortInput = {
  fields?: Maybe<Array<Maybe<MdxFieldsEnum>>>
  order?: Maybe<Array<Maybe<SortOrderEnum>>>
}

export type MdxWordCount = {
  paragraphs?: Maybe<Scalars["Int"]>
  sentences?: Maybe<Scalars["Int"]>
  words?: Maybe<Scalars["Int"]>
}

export type MdxWordCountFilterInput = {
  paragraphs?: Maybe<IntQueryOperatorInput>
  sentences?: Maybe<IntQueryOperatorInput>
  words?: Maybe<IntQueryOperatorInput>
}

/** Node Interface */
export type Node = {
  id: Scalars["ID"]
  parent?: Maybe<Node>
  children: Array<Node>
  internal: Internal
}
// ...
export type BlogIndexQuery = {
  allMdx: {
    nodes: Array<
      Pick<Mdx, "slug"> & {
        frontmatter?: Maybe<Pick<MdxFrontmatter, "title" | "date">>
      }
    >
  }
}